Move modules to subpackage modules.
