+++ nengo_spa/__init__.py
@@ -1,17 +1,17 @@
-from .actions import Actions
-from .assoc_mem import AssociativeMemory, ThresholdingAssocMem, WTAAssocMem
-from .basalganglia import BasalGanglia
-from .bind import Bind
-from .compare import Compare
-from .input import Input
-from .network import Network
-from .pointer import SemanticPointer
-from .product import Product
-from .scalar import Scalar
-from .state import State
-from .thalamus import Thalamus
-from .utils import similarity
-from .vocab import Vocabulary, VocabularyMap
+from nengo_spa.actions import Actions
+from nengo_spa.modules.assoc_mem import AssociativeMemory, ThresholdingAssocMem, WTAAssocMem
+from nengo_spa.modules.basalganglia import BasalGanglia
+from nengo_spa.modules.bind import Bind
+from nengo_spa.modules.compare import Compare
+from nengo_spa.modules.input import Input
+from nengo_spa.network import Network
+from nengo_spa.pointer import SemanticPointer
+from nengo_spa.modules.product import Product
+from nengo_spa.modules.scalar import Scalar
+from nengo_spa.modules.state import State
+from nengo_spa.modules.thalamus import Thalamus
+from nengo_spa.utils import similarity
+from nengo_spa.vocab import Vocabulary, VocabularyMap
 
 from nengo_spa import networks
 from nengo_spa import version
