+++ nengo_spa/actions.py
@@ -1,9 +1,8 @@
 """Parsing of SPA actions."""
-
 from nengo.exceptions import NetworkContextError, SpaParseError
 from nengo.network import Network
-from nengo_spa.basalganglia import BasalGanglia
-from nengo_spa.thalamus import Thalamus
+from nengo_spa.modules.basalganglia import BasalGanglia
+from nengo_spa.modules.thalamus import Thalamus
 from nengo_spa.spa_ast import (
     Action, ConstructionContext, DotProduct, Effect, Effects, Module,
     Reinterpret, Sink, Symbol, Translate)
