+++ nengo_spa/modules/assoc_mem.py
