+++ nengo_spa/modules/compare.py
