+++ nengo_spa/modules/input.py
