+++ nengo_spa/modules/product.py
