+++ nengo_spa/modules/scalar.py
