+++ nengo_spa/modules/state.py
