+++ nengo_spa/modules/tests/test_assoc_mem.py
@@ -2,7 +2,7 @@ import nengo
 import numpy as np
 import nengo_spa as spa
 from nengo_spa import Vocabulary
-from nengo_spa.assoc_mem import ThresholdingAssocMem, WTAAssocMem
+from nengo_spa.modules.assoc_mem import ThresholdingAssocMem, WTAAssocMem
 from nengo_spa.utils import similarity
