+++ nengo_spa/modules/tests/test_basalganglia.py
