+++ nengo_spa/modules/tests/test_bind.py
