+++ nengo_spa/modules/tests/test_compare.py
