+++ nengo_spa/modules/tests/test_cortical.py
