+++ nengo_spa/modules/tests/test_input.py
