+++ nengo_spa/modules/tests/test_state.py
