+++ nengo_spa/modules/tests/test_thalamus.py
